// typechain-types/AlphaERC721.ts
export declare namespace AlphaERC721 {
  export type ConstructorStruct = { name: string; symbol: string };

  export type ConstructorStructOutput = [string, string] & {
    name: string;
    symbol: string;
  };
}